refactor: 次のリファクタリングを実施 - 購入ボタン一覧のコンポーネント化 - お会計テーブルのコンポーネント化 - UIとロジックのディレクトリ分離
